+++ js/main.mjs
@@ -1,0 +1,3 @@
+import {generateObj} from './data.mjs';
+console.log(generateObj(2))
+
